+++ Configuration/config.properties
@@ -1,6 +1,6 @@
-baseURL = https://www.amazon.in/
+baseURL=https://www.amazon.in/
 chromepath=./Drivers/chromedriver
 firefoxpath=./Drivers/geckodriver
-browser = chrome
-runType = local
-homePageTitle = Online Shopping site in India: Shop Online for Mobiles, Books, Watches, Shoes and More - Amazon.in
+browser=chrome
+runType=local
+homePageTitle=Online Shopping site in India: Shop Online for Mobiles, Books, Watches, Shoes and More - Amazon.in
