+++ TestNG.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE suite SYSTEM "http://testng.org/testng-1.0.dtd">
-<suite name="ClipBoardHealthProject" >
+<suite name="ClipBoardHealthProject">
     <listeners>
         <listener class-name="com.ClipBoardHealth.utilities.Reporting"/>
     </listeners>
@@ -10,4 +10,4 @@
         </classes>
     </test>
 
-</suite> <!-- Suite -->
+</suite>
