+++ application.log
@@ -1,23 +1,24 @@
-2022-09-20 14:57:10,047  INFO [main] ReadConfig:29 - *******Url Launched******
-2022-09-20 14:57:10,051  INFO [main] ReadConfig:29 - *******Url Launched******
-2022-09-20 14:57:15,144  INFO [main] ClipBoardHealth  Project:53 - ***********Opened up the  browser***********
-2022-09-20 14:57:16,787  INFO [main] ReadConfig:29 - *******Url Launched******
-2022-09-20 14:57:16,792  INFO [main] ReadConfig:29 - *******Url Launched******
-2022-09-20 14:57:16,824  INFO [main] ReadConfig:29 - *******Url Launched******
-2022-09-20 14:57:16,834  INFO [main] ReadConfig:29 - *******Url Launched******
-2022-09-20 14:57:16,836  INFO [main] ReadConfig:29 - *******Url Launched******
-2022-09-20 14:57:16,836  INFO [main] ReadConfig:29 - *******Url Launched******
-2022-09-20 14:57:18,634  INFO [main] ReadConfig:29 - *******Url Launched******
-2022-09-20 14:57:18,634  INFO [main] ReadConfig:29 - *******Url Launched******
-2022-09-20 14:57:18,635  INFO [main] ReadConfig:29 - *******Url Launched******
-2022-09-20 14:57:18,635  INFO [main] ReadConfig:29 - *******Url Launched******
-2022-09-20 14:57:18,636  INFO [main] ReadConfig:29 - *******Url Launched******
-2022-09-20 14:57:18,636  INFO [main] ReadConfig:29 - *******Url Launched******
-2022-09-20 14:57:20,585  INFO [main] ReadConfig:29 - *******Url Launched******
-2022-09-20 14:57:20,586  INFO [main] ReadConfig:29 - *******Url Launched******
-2022-09-20 14:57:20,586  INFO [main] ReadConfig:29 - *******Url Launched******
-2022-09-20 14:57:20,586  INFO [main] ReadConfig:29 - *******Url Launched******
-2022-09-20 14:57:20,587  INFO [main] ReadConfig:29 - *******Url Launched******
-2022-09-20 14:57:20,587  INFO [main] ReadConfig:29 - *******Url Launched******
-2022-09-20 14:57:34,093  INFO [main] DriverSetUp:30 - [Resolution : 4K Ultra HD (3840 x 2160) Resolution || Refresh Rate : 100 Hertz, Connectivity: 4 HDMI ports to connect set top box || Blu-ray speakers or a gaming console || 2 USB ports to connect hard drives or other USB devices, Sound: 60 Watts Output - 4.2.2 Ch|| Powerful Speakers with Dolby Atmos || Surround Sound || Active Voice Amplifier || Adaptive Sound+, Smart TV Features : Mirroring || Tap View ||Multi-View || Music Wall || Mobile Camera Support || Wireless Dex || Auto Game Mode (ALLM) || Game Motion Plus || Dynamic Black EQ || Surround Sound || Super Ultra Wide Game View ||Mini Map Zoom || FreeSync Premium Pro || HGiG, Display: Matte Display || One Billion Color || PQI 4600 || 100% Colour Volume with Quantum Dot || Wide Viewing Angle || AI Upscale || Neo Quantum Processor 4K || Quantum Matrix Technology || Motion Xcelerator Turbo Pro || Real Depth Enhancer || LED Clear Motion, Warranty Information: 1 year comprehensive warranty plus additional 1 years on panel by brand from date of invoice, Installation: TV Table stand is not included in the box with this model. customer may ask for Table Top Stand or Wall Mount which will be provided to the customer at the time of installation, please directly call Samsung support on [Please visit brand website for tollfree numbers] and provide product's model name as well as seller's details mentioned on the invoice, Easy Returns: This product is eligible for replacement within 10 days of delivery in case of any product defects, damage or features not matching the description]
-2022-09-20 14:57:34,666  INFO [main] DriverSetUp:61 - browser closed
+2022-09-20 15:08:42,905  INFO [main] ReadConfig:29 - *******Url Launched******
+2022-09-20 15:08:42,909  INFO [main] ReadConfig:29 - *******Url Launched******
+2022-09-20 15:08:45,667  INFO [main] ClipBoardHealth  Project:50 - ***********Opened up the  browser***********
+2022-09-20 15:08:47,561  INFO [main] ReadConfig:29 - *******Url Launched******
+2022-09-20 15:08:47,573  INFO [main] ReadConfig:29 - *******Url Launched******
+2022-09-20 15:08:47,601  INFO [main] ReadConfig:29 - *******Url Launched******
+2022-09-20 15:08:47,631  INFO [main] ReadConfig:29 - *******Url Launched******
+2022-09-20 15:08:47,633  INFO [main] ReadConfig:29 - *******Url Launched******
+2022-09-20 15:08:47,634  INFO [main] ReadConfig:29 - *******Url Launched******
+2022-09-20 15:08:47,697  INFO [main] DriverSetUp:39 - Home Page Title Verified
+2022-09-20 15:08:47,712  INFO [main] ReadConfig:29 - *******Url Launched******
+2022-09-20 15:08:47,720  INFO [main] ReadConfig:29 - *******Url Launched******
+2022-09-20 15:08:47,722  INFO [main] ReadConfig:29 - *******Url Launched******
+2022-09-20 15:08:47,727  INFO [main] ReadConfig:29 - *******Url Launched******
+2022-09-20 15:08:47,746  INFO [main] ReadConfig:29 - *******Url Launched******
+2022-09-20 15:08:47,766  INFO [main] ReadConfig:29 - *******Url Launched******
+2022-09-20 15:08:49,208  INFO [main] ReadConfig:29 - *******Url Launched******
+2022-09-20 15:08:49,209  INFO [main] ReadConfig:29 - *******Url Launched******
+2022-09-20 15:08:49,210  INFO [main] ReadConfig:29 - *******Url Launched******
+2022-09-20 15:08:49,210  INFO [main] ReadConfig:29 - *******Url Launched******
+2022-09-20 15:08:49,211  INFO [main] ReadConfig:29 - *******Url Launched******
+2022-09-20 15:08:49,211  INFO [main] ReadConfig:29 - *******Url Launched******
+2022-09-20 15:09:01,935  INFO [main] DriverSetUp:30 - [Resolution : 4K Ultra HD (3840 x 2160) Resolution || Refresh Rate : 100 Hertz, Connectivity: 4 HDMI ports to connect set top box || Blu-ray speakers or a gaming console || 2 USB ports to connect hard drives or other USB devices, Sound: 60 Watts Output - 4.2.2 Ch|| Powerful Speakers with Dolby Atmos || Surround Sound || Active Voice Amplifier || Adaptive Sound+, Smart TV Features : Mirroring || Tap View ||Multi-View || Music Wall || Mobile Camera Support || Wireless Dex || Auto Game Mode (ALLM) || Game Motion Plus || Dynamic Black EQ || Surround Sound || Super Ultra Wide Game View ||Mini Map Zoom || FreeSync Premium Pro || HGiG, Display: Matte Display || One Billion Color || PQI 4600 || 100% Colour Volume with Quantum Dot || Wide Viewing Angle || AI Upscale || Neo Quantum Processor 4K || Quantum Matrix Technology || Motion Xcelerator Turbo Pro || Real Depth Enhancer || LED Clear Motion, Warranty Information: 1 year comprehensive warranty plus additional 1 years on panel by brand from date of invoice, Installation: TV Table stand is not included in the box with this model. customer may ask for Table Top Stand or Wall Mount which will be provided to the customer at the time of installation, please directly call Samsung support on [Please visit brand website for tollfree numbers] and provide product's model name as well as seller's details mentioned on the invoice, Easy Returns: This product is eligible for replacement within 10 days of delivery in case of any product defects, damage or features not matching the description]
+2022-09-20 15:09:02,372  INFO [main] DriverSetUp:58 - browser closed
