+++ src/main/java/com/ClipBoardHealth/pages/AllPageObjects.java
@@ -1,18 +1,18 @@
 package com.ClipBoardHealth.pages;
 
-import com.ClipBoardHealth.webPages.ProductPage;
 import com.ClipBoardHealth.webPages.HomePage;
-import com.ClipBoardHealth.webPages.ProductDetailsWebPage;
+import com.ClipBoardHealth.webPages.ProductDetailsPage;
+import com.ClipBoardHealth.webPages.ProductPage;
 
 public class AllPageObjects {
     public HomePage homePage;
-    public ProductDetailsWebPage listOfProductsPage;
+    public ProductDetailsPage listOfProductsPage;
 
     public ProductPage differentProductsPage;
 
     public AllPageObjects() {
         homePage = new HomePage();
-        listOfProductsPage = new ProductDetailsWebPage();
+        listOfProductsPage = new ProductDetailsPage();
         differentProductsPage = new ProductPage();
     }
 }
