+++ src/main/java/com/ClipBoardHealth/utilities/WebEventListener.java
@@ -1,6 +1,5 @@
 package com.ClipBoardHealth.utilities;
 
-import com.ClipBoardHealth.Base.BaseClass;
 import org.openqa.selenium.By;
 import org.openqa.selenium.OutputType;
 import org.openqa.selenium.WebDriver;
